So many tooltips, balances, QoL changes, abd bug fixes

+++ src/main/java/com/ombremoon/spellbound/common/world/entity/spell/Fireball.java
@@ -16,7 +16,7 @@ import net.minecraft.world.level.Level;
 import net.minecraft.world.phys.Vec3;
 
 public class Fireball extends VFXSpellProjectile<FireballSpell> {
-    private static final EntityDataAccessor<Integer> SIZE = SynchedEntityData.defineId(Fireball.class, EntityDataSerializers.INT);
+    private static final EntityDataAccessor<Float> SIZE = SynchedEntityData.defineId(Fireball.class, EntityDataSerializers.FLOAT);
     private static final EntityDataAccessor<Integer> STICK_TARGET = SynchedEntityData.defineId(Fireball.class, EntityDataSerializers.INT);
     private static final EntityDataAccessor<Boolean> STICKY = SynchedEntityData.defineId(Fireball.class, EntityDataSerializers.BOOLEAN);
     private int stickTime = 0;
@@ -38,7 +38,7 @@ public class Fireball extends VFXSpellProjectile<FireballSpell> {
     @Override
     protected void defineSynchedData(SynchedEntityData.Builder builder) {
         super.defineSynchedData(builder);
-        builder.define(SIZE, 0);
+        builder.define(SIZE, 0F);
         builder.define(STICKY, false);
         builder.define(STICK_TARGET, -1);
     }
@@ -78,11 +78,11 @@ public class Fireball extends VFXSpellProjectile<FireballSpell> {
         return EntityDimensions.scalable((this.getSize() * 0.5F) + 0.5F, (this.getSize() * 0.5F) + 0.5F);
     }
 
-    public int getSize() {
+    public float getSize() {
         return this.entityData.get(SIZE);
     }
 
-    public void setSize(int size) {
+    public void setSize(float size) {
         this.entityData.set(SIZE, size);
     }
